TelegramBot/requirements.txt
python-telegram-bot==20.3
tensorflow
librosa
numpy
nest_asyncio
